+++ CHANGELOG.md
@@ -16,6 +16,10 @@ Types of Changes
 
 ## [Unreleased]
 
+### Fixed
+
+- Addressed build issues.
+
 ## [0.1.0] - 2024-06-14
 
 ### Added